# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Roisfaozi_queue-base_internal_modules_counter_model.CreateCounterRequest:
    properties:
      branch_id:
        type: string
      code:
        maxLength: 50
        minLength: 2
        type: string
      name:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - branch_id
    - code
    - name
    type: object
  github_com_Roisfaozi_queue-base_internal_modules_counter_model.UpdateCounterRequest:
    properties:
      code:
        maxLength: 50
        minLength: 2
        type: string
      name:
        maxLength: 255
        minLength: 3
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    type: object
...
  github_com_Roisfaozi_queue-base_internal_modules_organization_model.CreateBranchRequest:
    properties:
      code:
        maxLength: 50
        minLength: 2
        type: string
      name:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - code
    - name
    type: object
...
  github_com_Roisfaozi_queue-base_internal_modules_organization_model.UpdateBranchRequest:
    properties:
      code:
        maxLength: 50
        minLength: 2
        type: string
      name:
        maxLength: 255
        minLength: 3
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    type: object
...
  github_com_Roisfaozi_queue-base_pkg_response.PageMetadata:
    properties:
      limit:
        type: integer
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
      total_item:
        type: integer
      total_page:
        type: integer
    type: object
...
  github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper:
    properties:
      data: {}
      paging:
        $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.PageMetadata'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a clean and modular boilerplate for Go REST APIs with RBAC,
    Audit Logs, and WebSockets.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Clean Boilerplate API
  version: "1.0"
paths:
...
  /branches:
    get:
      description: Returns all branches under active tenant context.
      parameters:
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: List branches
      tags:
      - branches
    post:
      consumes:
      - application/json
      description: Creates branch under active tenant context.
      parameters:
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      - description: Create Branch Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Roisfaozi_queue-base_internal_modules_organization_model.CreateBranchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Create branch
      tags:
      - branches
  /branches/{id}:
    delete:
      description: Deletes branch under active tenant context.
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Delete branch
      tags:
      - branches
    get:
      description: Returns a branch only when it belongs to active tenant context.
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Get branch by ID
      tags:
      - branches
    put:
      consumes:
      - application/json
      description: Updates branch under active tenant context.
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      - description: Update Branch Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Roisfaozi_queue-base_internal_modules_organization_model.UpdateBranchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Update branch
      tags:
      - branches
...
  /counters:
    get:
      description: Returns all counters under active tenant scope.
      parameters:
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: List counters
      tags:
      - counters
    post:
      consumes:
      - application/json
      description: Creates a new counter under active tenant scope.
      parameters:
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      - description: Create Counter Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Roisfaozi_queue-base_internal_modules_counter_model.CreateCounterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Create counter
      tags:
      - counters
  /counters/{id}:
    delete:
      description: Deletes counter under active tenant scope.
      parameters:
      - description: Counter ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Delete counter
      tags:
      - counters
    get:
      description: Returns counter details under active tenant scope.
      parameters:
      - description: Counter ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Get counter by ID
      tags:
      - counters
    put:
      consumes:
      - application/json
      description: Updates counter fields under active tenant scope.
      parameters:
      - description: Counter ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      - description: Update Counter Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Roisfaozi_queue-base_internal_modules_counter_model.UpdateCounterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Update counter
      tags:
      - counters
...
securityDefinitions:
  BearerAuth:
    description: '"Type ''Bearer '' followed by a space and the access token."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"